refactor: split distributed model and mailboxes
This will allow the reuse of the mailboxes in other parts of the code.
