+++ optimum/tpu/distributed_model.py
@@ -1,9 +1,7 @@
 # ruff: noqa: E402
 import os
 from enum import Enum
-from typing import Dict
 
-import torch
 from loguru import logger
 
 
@@ -12,67 +10,16 @@ os.environ["PJRT_DEVICE"] = "TPU"
 import torch.multiprocessing as mp
 import torch_xla.core.xla_model as xm
 import torch_xla.distributed.xla_multiprocessing as xmp
-from transformers import PretrainedConfig
 
 from optimum.tpu.modeling import AutoModelForCausalLM
+
+from .xla_mp_comm import AgentMailbox, RootMailbox
 
 
 class ModelCommand(Enum):
     LEAVE = 0
     PREFILL = 1
     DECODE = 2
-
-
-class RootMailbox:
-    def __init__(self, manager: mp.Manager):
-        self.root_bell = manager.Event()
-        self.root_command = manager.list()
-        self.model_ready = manager.Event()
-        self.output_data = manager.Value(torch.Tensor, torch.tensor([]))
-        self.model_config = manager.Value(PretrainedConfig, None)
-
-    @property
-    def config(self):
-        while True:
-            config = self.model_config.get()
-            if config is not None:
-                return config
-
-    def send(self, command: ModelCommand, data: Dict = None):
-        # First wait until model is ready to receive commands
-        self.model_ready.wait()
-        self.model_ready.clear()
-
-        self.root_command[:] = [command, data]
-        self.root_bell.set()
-        # wait again until model is ready, meaning command has been processed
-        self.model_ready.wait()
-        ret = self.output_data.get()
-        return ret
-
-
-class AgentMailbox:
-    def __init__(self, root_mailbox: RootMailbox):
-        self.root_bell = root_mailbox.root_bell
-        self.root_command = root_mailbox.root_command
-        self.model_ready = root_mailbox.model_ready
-        self.output_data = root_mailbox.output_data
-        self.model_config = root_mailbox.model_config
-
-    def receive(self):
-        self.root_bell.wait()
-        self.root_bell.clear()
-        return self.root_command
-
-    def send(self, data: torch.Tensor):
-        # Data needs to be moved to CPU before setting it
-        self.output_data.set(data.cpu())
-
-    @property
-    def command_data(self):
-        command = self.root_command[0]
-        data = self.root_command[1]
-        return command, data
 
 
 def _mp_fn(rank, model_id, root_mailbox: RootMailbox, sample_fn: callable):
